Quick and dirty quilljs test
I tried to see if it was possible to add Quilljs to our page. It is using a package from NPM, but there seems to issues with getting asciimath displaying since the mathjax code lies in its own component. Decided to make this a separate branch for now until we can decide if it is worth implementing in the main branch.